requirements.txt
kafka-python
faust-streaming
aiokafka==0.11.0
dagster
dagster-webserver
dagster-aws
dagster-snowflake
pandas
polars
pyarrow
redis
boto3
requests
python-dotenv
deltalake
dbt-snowflake
dagster-dbt
prometheus-client
pytest
ruff
